Rename pioneer.py to drone.py
