+++ blueye/sdk/__init__.py
@@ -1,3 +1,3 @@
-from .pioneer import Drone, Pioneer
+from .drone import Drone, Pioneer
 from .utils import open_local_documentation
 from .constants import WaterDensities
